chore: bifurcate EnektoService into EnektoFormService
